+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * Triggering redraw at active scrap change by activeScrapID observable.
   * Removing duplicated code at MPAzimuthPickerWidget.
   * Changing THElement.parent() so thFile parameter is optional.
+  * Implemented TH2FileEditElementEditController.addOutdatedLineSegmentCloneMPID().
 
 ## 0.2.28 - 2025-12-19 - The Kichute and Conga release
 * New features: